call dispatch directly in component & move updateChildCheckbox out of reducer
